Use OR instead of AND for file extraction flags
The bitwise AND was effectively always evaluating to 0 and not passing along
flags when they're specified, e.g. from the [libarchive](https://github.com/chef-cookbooks/libarchive/blob/v2.0.0/libraries/helper.rb#L47)
cookbook.

As I read it, it looks like this should be an OR rather than an AND so the
desired flags are always set.

Signed-off-by: Jonathan Hartman <j@hartman.io>